+++ pom.xml
@@ -11,7 +11,7 @@
     <groupId>com.adrian.springframework</groupId>
     <artifactId>mvc-rest</artifactId>
     <version>0.0.1-SNAPSHOT</version>
-    <name>mvc-rest</name>
+    <name>webflux-rest</name>
     <description>Demo project for Spring Boot</description>
 
     <properties>
