+++ src/main/java/com/adrian/springframework/WebFluxRestApplication.java
@@ -4,10 +4,10 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class MvcRestApplication {
+public class WebFluxRestApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(MvcRestApplication.class, args);
+        SpringApplication.run(WebFluxRestApplication.class, args);
     }
 
 }
