+++ src/test/java/com/adrian/springframework/webfluxrest/WebFluxRestApplicationTests.java
@@ -1,10 +1,10 @@
-package com.adrian.springframework.mvcrest;
+package com.adrian.springframework.webfluxrest;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class MvcRestApplicationTests {
+class WebFluxRestApplicationTests {
 
     @Test
     void contextLoads() {
